organize SessionView
