+++ src/views/SessionView/Whiteboard.vue
@@ -750,26 +750,26 @@ canvas {
 }
 
 #eraseButton {
-  background-image: url('../../assets/eraser_icon.svg');
+  background-image: url('~src/assets/eraser_icon.svg');
 }
 
 #drawButton {
-  background-image: url('../../assets/pen_icon.svg');
+  background-image: url('~src/assets/pen_icon.svg');
 }
 
 #undoButton {
-  background-image: url('../../assets/undo_icon.svg');
+  background-image: url('~src/assets/undo_icon.svg');
 }
 
 #textButton {
-  background-image: url('../../assets/Aa.png');
+  background-image: url('~src/assets/Aa.png');
 }
 
 #clearButton {
-  background-image: url('../../assets/clear_icon.svg');
+  background-image: url('~src/assets/clear_icon.svg');
 }
 
 #openColorsButton {
-  background-image: url('../../assets/color_icon.svg');
+  background-image: url('~src/assets/color_icon.svg');
 }
 </style>
